Modified the rest of the printk's to use the LOG construct. Removed some unused code. Modified the getPacketCount and getByteCount to iterate the flows that have multiple entries in the NetFPGA (Wildcards) and total the individual entry values before returning.

+++ datapath/hwtable_nf2/hwtable_nf2.h
@@ -62,7 +62,7 @@ int nf2_get_table_type(struct sw_flow *flow);
 int nf2_build_and_write_flow(struct sw_flow *flow);
 void nf2_delete_private(void* private);
 
-uint32_t nf2_get_packet_count(struct net_device *dev, struct sw_flow_nf2 *sfw);
-uint32_t nf2_get_byte_count(struct net_device *dev, struct sw_flow_nf2 *sfw);
+uint64_t nf2_get_packet_count(struct net_device *dev, struct sw_flow_nf2 *sfw);
+uint64_t nf2_get_byte_count(struct net_device *dev, struct sw_flow_nf2 *sfw);
 
 #endif /*HWTABLENF2_H_*/
